No way s_h*artsPerDevice can be just uchars

+++ src/GlobalData.h
@@ -73,8 +73,8 @@ struct GlobalData {
 	uchar* 			s_hDeviceMap; // one uchar for each cell, tells  which device the cell has been assigned to
 
 	// counter: how many particles per device
-	uchar s_hPartsPerDevice[MAX_DEVICES_PER_CLUSTER]; // TODO: can change to PER_NODE if not compiling for multinode
-	uchar s_hStartPerDevice[MAX_DEVICES_PER_CLUSTER]; // ditto
+	uint s_hPartsPerDevice[MAX_DEVICES_PER_CLUSTER]; // TODO: can change to PER_NODE if not compiling for multinode
+	uint s_hStartPerDevice[MAX_DEVICES_PER_CLUSTER]; // ditto
 
 	// pinned memory var to retrieve dt asynchronously
 	//float *pin_maxcfl;
